requirements.txt
# Core
pandas>=2.0.0
numpy>=1.23.0
requests>=2.31.0

# OCR
pytesseract>=0.3.10
easyocr>=1.7.1
opencv-python>=4.8.0
Pillow>=9.5.0

# Google Search API
serpapi>=0.1.5

# Article extraction
newspaper3k>=0.2.8
beautifulsoup4>=4.12.2
readability-lxml>=0.8.1
lxml>=4.9.3

# NLP & text processing
nltk>=3.8.1
spacy>=3.6.1
rapidfuzz>=3.1.1

# Machine Learning
scikit-learn>=1.3.0

# Optional (useful utilities)
tqdm>=4.66.0        # progress bars
python-Levenshtein>=0.21.1  # (rapidfuzz is usually enough but good to have)
